Moved groupers; Updated documentation and made corrections based on Fred and Xinyu's comments
