+++ samza-core/src/main/java/org/apache/samza/config/JobCoordinatorConfig.java
@@ -22,8 +22,7 @@ package org.apache.samza.config;
 import com.google.common.base.Strings;
 
 public class JobCoordinatorConfig extends MapConfig {
-  // TODO: Change this to job-coordinator.factory
-  private static final String JOB_COORDINATOR_FACTORY = "job.coordinator.factory";
+  public static final String JOB_COORDINATOR_FACTORY = "job-coordinator.factory";
 
   public JobCoordinatorConfig(Config config) {
     super(config);
