+++ samza-core/src/main/java/org/apache/samza/container/grouper/task/SingleContainerGrouperFactory.java
@@ -17,12 +17,11 @@
  * under the License.
  */
 
-package org.apache.samza.test.processor;
+package org.apache.samza.container.grouper.task;
 
 import org.apache.samza.config.Config;
+import org.apache.samza.config.JobConfig;
 import org.apache.samza.container.TaskName;
-import org.apache.samza.container.grouper.task.TaskNameGrouper;
-import org.apache.samza.container.grouper.task.TaskNameGrouperFactory;
 import org.apache.samza.job.model.ContainerModel;
 import org.apache.samza.job.model.TaskModel;
 
@@ -32,15 +31,13 @@ import java.util.Map;
 import java.util.Set;
 
 public class SingleContainerGrouperFactory implements TaskNameGrouperFactory {
-  private static final String PROCESSOR_ID = "processor.id";
-
   @Override
   public TaskNameGrouper build(Config config) {
-    return new SingleContainerGrouper(config.getInt(PROCESSOR_ID));
+    return new SingleContainerGrouper(config.getInt(JobConfig.PROCESSOR_ID()));
   }
 }
 
-class SingleContainerGrouper implements TaskNameGrouper  {
+class SingleContainerGrouper implements TaskNameGrouper {
   private final int containerId;
 
   SingleContainerGrouper(int containerId) {
